[Misc] Move models signals
