+++ apps/track/models/__init__.py
@@ -1,4 +1,20 @@
+import gpxpy
+from django.db.models.signals import post_save
+from django.dispatch import receiver
+
 from .comment import Comment
 from .like import Like
 from .track import Track
 from .track_stat import TrackStat
+
+
+@receiver(post_save, sender=Track)
+def track_post_save(sender, instance: Track, created: bool, *args, **kwargs):
+    if not TrackStat.objects.filter(track=instance).exists():
+        track_stat = TrackStat(track=instance)
+        track_stat.compute()
+        track_stat.save()
+    if created:
+        gpx = gpxpy.parse(instance.source_file.open())
+        instance.datetime = gpx.get_time_bounds().start_time
+        instance.save()
